move function and bump version
- moves associateJoin to Database.Esqueleto
- relaxes bounds on containers dep
-
