+++ changelog.md
@@ -1,3 +1,9 @@
+3.1.2
+========
+
+- @tippenein
+  - [#149](https://github.com/bitemyapp/esqueleto/pull/157): Added `associateJoin` query helpers.
+
 3.1.1
 ========
